Refactor Unit test and move data into fixture
